+++ auth_config.json
@@ -1,0 +1,4 @@
+{
+  "domain": "dev-3k36-3cg.us.auth0.com",
+  "clientId": "AUMp2619jwCfVNLdeW6uRSTJUt3wQk01"
+}
