+++ src/app/app.component.html
@@ -2,11 +2,11 @@
 <!-- Views render in here -->
 <nav>
     <ul>
-        
+
             <a routerLink="/home"
             routerLinkActive="active"
             ariaCurrentWhenActive="page"><img src = "assets/court_logo.jpeg" alt = "court_logo"></a>
-        
+
         <li>
             <a routerLink="/home"
             routerLinkActive="active"
