+++ src/app/components/login/login.component.html
@@ -1,0 +1,1 @@
+<button class="btn btn-primary btn-block" (click)="loginWithRedirect()">Login</button>
