+++ src/app/components/login/login.component.ts
@@ -1,0 +1,18 @@
+import { Component, OnInit } from '@angular/core';
+import { AuthService } from '@auth0/auth0-angular';
+
+@Component({
+  selector: 'app-login',
+  templateUrl: './login.component.html',
+})
+export class LoginComponent implements OnInit {
+
+  constructor(public auth: AuthService) { }
+
+  ngOnInit(): void {
+  }
+
+  loginWithRedirect(): void {
+    this.auth.loginWithRedirect();
+  }
+}
