--- src/app/login/login.component.html
@@ -1,19 +1,0 @@
-<div class="container">
-  <form>
-    <div class="form-group">
-      <label for = "username">Username:</label>
-      <input type="text" class="form-control" id="username">
-    </div>
-    <div class="form-group">
-      <label for="email">Email address:</label>
-      <input type="email" class="form-control" id="email">
-    </div>
-    <div class="form-group">
-      <label for="pwd">Password:</label>
-      <input type="password" class="form-control" id="pwd">
-    </div>
-    <div class="checkbox">
-      <label><input type="checkbox"> Remember me</label>
-    </div>
-    <button mat-raised button color="primary" type="submit" class="btn btn-default">Submit</button>
-  </form>
