+++ src/environments/auth0.prod.ts
@@ -1,0 +1,7 @@
+export const auth0 = {
+  auth: {
+    domain: "dev-3k36-3cg.us.auth0.com",
+    clientId: "AUMp2619jwCfVNLdeW6uRSTJUt3wQk01",
+    redirectUri: window.location.origin,
+  },
+};
