+++ src/environments/auth0.ts
@@ -1,0 +1,8 @@
+import { domain, clientId } from '../../auth_config.json';
+export const auth0 = {
+  auth: {
+    domain,
+    clientId,
+    redirectUri: window.location.origin,
+  },
+};
